Refactor pagination implementaiton
Given that the new StatusCake SDK implementations are well documented
and follow a set standard, we can assume that the pagination
requirements do not differ between endpoint calls.

This means that we can refactor the code so that pagination is handled
by a single private method.

The method will take a function definition and a set of args, perform
the required pagination and return a response to the caller.